requirements.txt
scipy 
numpy
matplotlib
ipython
pyqt
h5py
